Move package reclaim options to one controller
