+++ LBH.AdultSocialCare.Api/V1/Controllers/HomeCare/HomeCarePackageReclaimController.cs
@@ -1,15 +1,10 @@
-using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Boundary.HomeCarePackageReclaimBoundary.Request;
 using LBH.AdultSocialCare.Api.V1.Boundary.HomeCarePackageReclaimBoundary.Response;
-using LBH.AdultSocialCare.Api.V1.Boundary.PackageReclaimsBoundary.Response;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.UseCase.HomeCarePackageReclaimUseCase.Interfaces;
-using LBH.AdultSocialCare.Api.V1.UseCase.ReclaimUseCase.Interfaces;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
 
 namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
 {
@@ -21,19 +16,10 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
     public class HomeCarePackageReclaimController : BaseController
     {
         private readonly ICreateHomeCarePackageReclaimUseCase _createHomeCarePackageReclaimUseCase;
-        private readonly IGetAllAmountOptionUseCase _getAllAmountOptionUseCase;
-        private readonly IGetAllReclaimCategoryUseCase _getAllReclaimCategoryUseCase;
-        private readonly IGetAllReclaimFromUseCase _getAllReclaimFromUseCase;
 
-        public HomeCarePackageReclaimController(ICreateHomeCarePackageReclaimUseCase createHomeCarePackageReclaimUseCase,
-            IGetAllAmountOptionUseCase getAllAmountOptionUseCase,
-            IGetAllReclaimCategoryUseCase getAllReclaimCategoryUseCase,
-            IGetAllReclaimFromUseCase getAllHomeCareReclaimFromUseCase)
+        public HomeCarePackageReclaimController(ICreateHomeCarePackageReclaimUseCase createHomeCarePackageReclaimUseCase)
         {
             _createHomeCarePackageReclaimUseCase = createHomeCarePackageReclaimUseCase;
-            _getAllAmountOptionUseCase = getAllAmountOptionUseCase;
-            _getAllReclaimCategoryUseCase = getAllReclaimCategoryUseCase;
-            _getAllReclaimFromUseCase = getAllHomeCareReclaimFromUseCase;
         }
 
         [ProducesResponseType(typeof(HomeCarePackageClaimResponse), StatusCodes.Status200OK)]
@@ -59,36 +45,6 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
             var homeCarePackageClaimResponse =
                 await _createHomeCarePackageReclaimUseCase.ExecuteAsync(homeCarePackageClaimForCreationDomain).ConfigureAwait(false);
             return Ok(homeCarePackageClaimResponse);
-        }
-
-        [ProducesResponseType(typeof(IEnumerable<ReclaimAmountOptionResponse>), StatusCodes.Status200OK)]
-        [ProducesDefaultResponseType]
-        [HttpGet]
-        [Route("amount-option")]
-        public async Task<ActionResult<IEnumerable<ReclaimAmountOptionResponse>>> GetAmountOptionsList()
-        {
-            var result = await _getAllAmountOptionUseCase.GetAllAsync().ConfigureAwait(false);
-            return Ok(result);
-        }
-
-        [ProducesResponseType(typeof(IEnumerable<ReclaimFromResponse>), StatusCodes.Status200OK)]
-        [ProducesDefaultResponseType]
-        [HttpGet]
-        [Route("reclaim-from")]
-        public async Task<ActionResult<IEnumerable<ReclaimFromResponse>>> GetTypeOfReclaimFromOptionList()
-        {
-            var result = await _getAllReclaimFromUseCase.GetAllAsync().ConfigureAwait(false);
-            return Ok(result);
-        }
-
-        [ProducesResponseType(typeof(IEnumerable<ReclaimCategoryResponse>), StatusCodes.Status200OK)]
-        [ProducesDefaultResponseType]
-        [HttpGet]
-        [Route("reclaim-category")]
-        public async Task<ActionResult<IEnumerable<ReclaimCategoryResponse>>> GetTypeOfReclaimCategoryOptionList()
-        {
-            var result = await _getAllReclaimCategoryUseCase.GetAllAsync().ConfigureAwait(false);
-            return Ok(result);
         }
     }
 }
